==> 08_lightfm_annoy_funds.ipynb <==
算法抄襲自:
[github repo](https://github.com/lesterlitch/misc/blob/master/Light%2Bfm%2Bannoy%2Band%2Bproduct%2Bsearch%2Bexample.ipynb)

# Lightfm

from scipy import sparse as sp
from lightfm import LightFM
from lightfm.evaluation import precision_at_k,recall_at_k,auc_score
import copy
import itertools
import numpy as np 
import lightfm
import pickle 
import pandas as pd 

#fundid_names_df.to_csv('./funds/fundid_to_name.csv',index=False)

with open('./funds-dataset/sp_funds_datasets.pickle','rb') as f:
    data = pickle.load(f)
    
test = data['test']
train = data['train']
user_idxs = data['user_idxs']
idx_to_userid = data['idx_to_userid']
userid_to_idx = data['userid_to_idx']
idx_to_itemid = data['idx_to_itemid']
itemid_to_idx = data['itemid_to_idx']

fundid_names_df = pd.read_csv('./funds-dataset/fundid_to_name.csv',encoding='cp950')
fundid_to_names = {}

for d in fundid_names_df.to_dict('records'):
    fundid_to_names[d['基金代碼']] = d['基金中文名稱']
#%% 

model = LightFM(learning_rate=0.01, loss='warp')
model.fit(train, epochs=10)

train_precision = precision_at_k(model, train, k=10).mean()
test_precision = precision_at_k(model, test, k=10).mean()
train_recall = recall_at_k(model,train,k=10).mean()
test_recall = recall_at_k(model,test,k=10).mean()

train_auc = auc_score(model, train).mean()
test_auc = auc_score(model, test).mean()
print('Recall: train {:.2f}%, test {:.2f}%'.format(100*train_recall,100*test_recall))
print('Precision: train {:.2f}% , test {:.2f}%.'.format(100*train_precision, 100*test_precision))
print('AUC: train {:.2f}, test {:.2f}.'.format(train_auc, test_auc))

def sample_recommendation_original(model, data, user_ids, print_output=True):
    
    train = data['train']
    test = data['test']
    assert isinstance(train,sp.csr_matrix) and isinstance(test,sp.csr_matrix)
        
    n_users, n_items = train.shape

    for user_id in user_ids:
        
        known_positives_itemids = [ 
                idx_to_itemid[e] for e in train[user_id].indices
                ]
        known_positives_item_names = [
                fundid_to_names[e] for e in known_positives_itemids
                ]
        scores = model.predict(user_id, np.arange(n_items))
        top_items_ids = [idx_to_itemid[e] for e in np.argsort(-scores)]
        if print_output == True:
            print("User %s" % user_id)
            print("     Known positives:")

            for x in known_positives_item_names[:3]:
                print("        %s" % x)

            print("     Recommended:")

            for x in top_items_ids[:3]:
                print("        %s" % fundid_to_names[x])

sample_recommendation_original(model,data,range(3))

___

# SVD

from scipy.sparse.linalg import svds

train.asfptype()

user_factors, _, item_factors =svds(train.asfptype(),k=50)

class TopRelated:
    def __init__(self, artist_factors):
        # fully normalize artist_factors, so can compare with only the dot product
        norms = np.linalg.norm(artist_factors, axis=-1)
        self.factors = artist_factors / norms[:, np.newaxis]

    def get_related(self, artistid, N=10):
        scores = self.factors.dot(self.factors[artistid]) # cosine 
        best = np.argpartition(scores, -N)[-N:]
        return sorted(zip(best, scores[best]), key=lambda x: -x[1])

TopRelated(user_factors).get_related(1)

### ANN 

class ApproximateTopRelated(object):
    def __init__(self, artist_factors, treecount=20):
        index = annoy.AnnoyIndex(artist_factors.shape[1], 'angular')
        for i, row in enumerate(artist_factors):
            index.add_item(i, row)
        index.build(treecount)
        self.index = index

    def get_related(self, artistid, N=10):
        neighbours = self.index.get_nns_by_item(artistid, N)
        return sorted(((other, 1 - self.index.get_distance(artistid, other))
                      for other in neighbours), key=lambda x: -x[1])

u_app = ApproximateTopRelated(user_factors)
u_app.get_related(1)

### pair-wise distance

from sklearn.metrics.pairwise import pairwise_distances

p_dist = pairwise_distances(user_factors)

np.argpartition(p_dist[0],-10)

f = lambda user: np.argpartition(p_dist[user],10)[:10]


print(f(1)) ## closet users
p_dist[1,f(1)]

_____

# Lightfm + ANN

Now we'll do an Annoy example showing how we can do user recommendations using a neat trick outlined by the 
[Xbox recomendations team]( https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/XboxInnerProduct.pdf)

def sample_recommendation_annoy(model,data,user_ids,print_output=True):
    train = data['train']
    test = data['test']
    n_users, n_items = train.shape

    for user_id in user_ids:
        
        known_positives_itemids = [ 
                idx_to_itemid[e] for e in train[user_id].indices
                ]
        known_positives_item_names = [
                fundid_to_names[e] for e in known_positives_itemids
                ]
        
        top_items_ids = [idx_to_itemid[e] for e in t_member.get_nns_by_vector(
            np.append(user_vectors[user_id], 0), 50)]
        
        if print_output == True:
            print("User %s" % user_id)
            print("     Known positives:")

            for x in known_positives_item_names[:3]:
                print("        %s" % x)

            print("     Recommended:")

            for x in top_items_ids[:3]:
                print("        %s" % fundid_to_names[x])

sample_recommendation_annoy(model,data,user_ids=[0,1,2])

import annoy 
from annoy import AnnoyIndex

item_vectors = model.item_embeddings

item_vectors.shape

np.sqrt(max_norm)

norm_data.shape

norms = np.linalg.norm(item_vectors, axis=-1)
# self.factors = artist_factors / norms[:, numpy.newaxis]

norms[:,np.newaxis]

class ApproximateTopRelated(object):
    def __init__(self, artist_factors, treecount=20):
        index = annoy.AnnoyIndex(artist_factors.shape[1], 'angular')
        for i, row in enumerate(artist_factors):
            index.add_item(i, row)
        index.build(treecount)
        self.index = index

    def get_related(self, artistid, N=10):
        neighbours = self.index.get_nns_by_item(artistid, N)
        return sorted(((other, 1 - self.index.get_distance(artistid, other))
                      for other in neighbours), key=lambda x: -x[1])

approx_related = ApproximateTopRelated(item_vectors)
approx_related.get_related(10)

norms = np.linalg.norm(item_vectors, axis=1)
max_norm = norms.max()
extra_dimension = np.sqrt(max_norm ** 2 - norms ** 2)
norm_data = np.append(
    item_vectors, extra_dimension.reshape(norms.shape[0], 1), axis=1)


f_member = norm_data.shape[1]
t_member = AnnoyIndex(f_member,'angular')  # Length of item vector that will be indexed

for i in range(norm_data.shape[0]):
    v = norm_data[i]
    t_member.add_item(i, v)

t_member.build(200)

user_id = 10
topn = 10 
t_member.get_nns_by_vector(np.append(user_vectors[user_id],0),topn)

recall, precision on annoy 

def eval_recommendation_test(model,test,topn=10,annoy=True):
    n_users,n_items = test.shape
    hr = 0
    n_test = 0
    for user_id in range(n_users):
        if test[user_id].indices:
            if annoy:
                top_items_idxs = t_member.get_nns_by_vector(np.append(user_vectors[user_id],0),topn)
            else:
                scores = model.predict(user_id, np.arange(n_items))
                top_items_idxs = np.argsort(-scores)[:topn]
                
            test_fundidx = test[user_id].indices
            hr += np.any(np.intersect1d(test_fundidx,top_items_idxs))
            n_test += 1
    return hr/n_test

%time rat10_annoy = eval_recommendation_test(model,test,annoy=True)
%time rat10 = eval_recommendation_test(model,test,annoy=False)

print('recall:{:.2f}% at topn=10, without approximation '.format(rat10*100))
print('recall:{:.2f}% at topn=10, with approximation '.format(rat10_annoy*100))

_____

def nearest_items_Annoy(itemid, index, n=10, print_output=True):
    nn = index.get_nns_by_item(itemid, 10)
    if print_output == True:
        print('Closest to %s : \n' % movielens['item_labels'][movie_id])
    titles = [movielens['item_labels'][i] for i in nn]
    if print_output == True:
        print("\n".join(titles))



f = item_vectors.shape[1]  # Length of item vector that will be indexed
t = AnnoyIndex(f)
for i in range(item_vectors.shape[0]):
    v = item_vectors[i]
    t.add_item(i, v)

t.build(10)  # 10 trees

itemid = 0
print('fund:{}'.format(fundid_to_names[idx_to_itemid[itemid]]))
[fundid_to_names[idx_to_itemid[e]] for e in t.get_nns_by_item(0,10)]
